packages/ml_core/pyproject.toml
[project]
name = "ml_core"
version = "0.0.1"
description = "Unified ML model interface and shared utilities for substation forecasting."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mlflow-skinny",
    "polars>=1.0.0",
    "patito",
    "pydantic>=2.0.0",
    "dagster",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ml_core"]

[tool.ruff]
line-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = 100
